move EC-MS calculations
