+++ src/ixdat/calculators/ecms_calculators.py
@@ -1,0 +1,167 @@
+# -*- coding: utf-8 -*-
+"""
+Created on Thu Aug  8 20:48:06 2024
+
+@author: Søren
+"""
+import warnings
+import numpy as np
+from scipy.optimize import minimize
+from ..config import plugins
+from ..measurements import Calculator
+from ..plotters.ms_plotter import STANDARD_COLORS
+from ..constants import FARADAY_CONSTANT
+from .ms_calculators import MSCalibration, MSCalResult
+
+
+class ECMSCalibration(Calculator):
+    """Class for calibrations done using ECMSMeasurements
+
+    Its classmethods return MSCalibration objects.
+    """
+
+    @classmethod
+    def ecms_calibration(cls, measurement, mol, mass, n_el, tspan, tspan_bg=None):
+        """Calibrate for mol and mass based on one period of steady electrolysis
+
+        Args:
+            measurement (ECMSMeasurement): measurement with calibration data
+            mol (str): Name of the molecule to calibrate
+            mass (str): Name of the mass at which to calibrate
+            n_el (str): Number of electrons passed per molecule produced (remember the
+                sign! e.g. +4 for O2 by OER and -2 for H2 by HER)
+            tspan (tspan): The timespan of steady electrolysis
+            tspan_bg (tspan): The time to use as a background
+
+        Return MSCalResult: The result of the ecms_calibration
+        """
+        if plugins.use_siq:
+            warnings.warn(
+                "spectro_inlets_quantification is active but you are using the native "
+                "ixdat version of `ECMSMeasurement.ecms_calibration`"
+            )
+        Y = measurement.integrate_signal(mass, tspan=tspan, tspan_bg=tspan_bg)
+        Q = measurement.integrate("raw_current", tspan=tspan) * 1e-3
+        n = Q / (n_el * FARADAY_CONSTANT)
+        F = Y / n
+        cal = MSCalResult(
+            name=f"{mol}@{mass}",
+            mol=mol,
+            mass=mass,
+            cal_type="ecms_calibration",
+            F=F,
+        )
+        return MSCalibration(ms_cal_results=[cal], measurement=measurement)
+
+    @classmethod
+    def ecms_calibration_curve(
+        cls,
+        measurement,
+        mol,
+        mass,
+        n_el,
+        tspan_list=None,
+        selector_name=None,
+        selector_list=None,
+        t_steady_pulse=None,
+        tspan_bg=None,
+        force_through_zero=False,
+        ax="new",
+        axes_measurement=None,
+        axes_measurement_J_name="raw_current",
+        return_ax=False,
+    ):
+        """Fit mol's sensitivity at mass based on steady periods of EC production.
+
+        Args:
+            mol (str): Name of the molecule to calibrate
+            mass (str): Name of the mass at which to calibrate
+            n_el (str): Number of electrons passed per molecule produced (remember the
+                sign! e.g. +4 for O2 by OER and -2 for H2 by HER)
+            tspan_list (list of tspan): The timespans of steady electrolysis
+            selector_name (str): Name of selector which identifies the periods
+                of steady electrolysis for automatic selection of timespans of steady
+                electrolysis. E.g. "selector" or "Ns" for biologic EC data
+            selector_list (list): List of values for selector_name for automatic
+                selection of timespans of steady electrolysis
+            t_steady_pulse (float): Length of steady electrolysis for each segment
+                given by selector_list. Defaults to None = entire length of segment
+            tspan_bg (tspan): The time to use as a background
+            force_through_zero (boolean): Whether to force the calibration curve through
+                zero. This can be done when confident in the background subtraction.
+            ax (Axis): The axis on which to plot the ms_calibration curve result.
+                Defaults to a new axis.
+            axes_measurement (list of Axes): The EC-MS plot axes to highlight the
+                ms_calibration on. Defaults to None. These axes are not returned.
+            axes_measurement_J_name (str): The J_name used in the axis passed
+                to axes_measurement. Must be passed manually as the axis does not "know"
+                its J_name. Defaults to "raw_current". IMPORTANT: the method still uses
+                "raw_current" to calculate the sensitivity factor, this J_name is only
+                used for plotting.
+            return_ax (bool): Whether to return the axis on which the calibration curve
+                is plotted together with the MSCalResult. Defaults to False.
+
+        Return MSCalResult(, Axis): The result of the ms_calibration (and calibration
+            curve axis if requested) based on integration of selected time periods.
+        """
+
+        axis_ms = axes_measurement[0] if axes_measurement else None
+        axis_current = axes_measurement[3] if axes_measurement else None
+        Y_list = []
+        n_list = []
+        if not tspan_list:
+            tspan_list = measurement._get_tspan_list(
+                selector_list, selector_name, t_steady_pulse
+            )
+        for tspan in tspan_list:
+            Y = measurement.integrate_signal(
+                mass, tspan=tspan, tspan_bg=tspan_bg, ax=axis_ms
+            )
+            # FIXME: plotting current by giving integrate() an axis doesn't work great.
+            if (
+                axes_measurement
+            ):  # FIXME: need to run twice, once to plot, once to calculate Q
+                measurement.integrate(
+                    axes_measurement_J_name, tspan=tspan, ax=axis_current
+                )
+            Q = measurement.integrate("raw_current", tspan=tspan)
+            Q *= 1e-3  # mC --> [C]
+            n = Q / (n_el * FARADAY_CONSTANT)
+            Y_list.append(Y)
+            n_list.append(n)
+        n_vec = np.array(n_list)
+        Y_vec = np.array(Y_list)
+        n_fit = np.array([0, max(n_vec)])
+        if force_through_zero:
+
+            def rms_error(F_guess):
+                return np.mean((Y_vec - F_guess * n_vec) ** 2)
+
+            F_guess_0 = np.sum(Y_vec) / np.sum(n_vec)
+            res = minimize(rms_error, F_guess_0)
+            F = res.x[0]
+            Y_fit = n_fit * F
+        else:
+            pfit = np.polyfit(n_vec, Y_vec, deg=1)
+            F = pfit[0]
+            Y_fit = n_fit * pfit[0] + pfit[1]
+
+        if ax:
+            color = STANDARD_COLORS[mass]
+            if ax == "new":
+                ax = measurement.plotter.new_ax()
+                ax.set_xlabel("amount produced / [nmol]")
+                ax.set_ylabel("integrated signal / [nC]")
+            ax.plot(n_vec * 1e9, Y_vec * 1e9, "o", color=color)
+            ax.plot(n_fit * 1e9, Y_fit * 1e9, "--", color=color)
+
+        cal = MSCalResult(
+            name=f"{mol}@{mass}",
+            mol=mol,
+            mass=mass,
+            cal_type="ecms_calibration_curve",
+            F=F,
+        )
+        return MSCalibration(ms_cal_results=[cal], measurement=measurement)
+
+    from_siq = MSCalibration.from_siq
